backend/pyproject.toml
[project]
name = "ai-chat-bot-backend"
version = "0.1.0"
description = "AI Chat Bot Demo - FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.5.0",
    "PyJWT[crypto]>=2.9.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "mcp>=1.1.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
